+++ app/sprinkles/core/src/Facades/Cache.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ */
+
+namespace UserFrosting\Sprinkle\Core\Facades;
+
+use UserFrosting\System\Facade;
+
+/**
+ * Implements facade for the "cache" service
+ *
+ * @author Louis Charette
+ */
+class Cache extends Facade
+{
+    /**
+     * Get the registered name of the component.
+     *
+     * @return string
+     */
+    protected static function getFacadeAccessor()
+    {
+        return 'cache';
+    }
+}